+++ tests/testthat/helper-env.R
@@ -1,1 +1,13 @@
-Sys.setenv(BIGQUERY_DATASET = "test")
+library(bigrquery)
+
+if (bqTokenFileValid()) {
+  bqAuth()
+  ds <- bq_test_dataset()
+  Sys.setenv(BIGQUERY_DATASET = ds$dataset)
+  Sys.setenv(BIGQUERY_PROJECT = Sys.getenv("BIGQUERY_TEST_PROJECT"))
+  if (!bqDatasetExists()) {
+    bqCreateDataset()
+  }
+}
+
+