+++ tests/testthat/partition-transform.sql
@@ -1,0 +1,6 @@
+SELECT
+  count * 2 as count
+FROM
+  %2$s
+WHERE
+  _PARTITIONTIME = TIMESTAMP('%1$s')