+++ tests/testthat/test.bq.table.R
@@ -1,35 +1,22 @@
 library(bigrquery)
 library(mockery)
 
-context("BigQuery Create Table")
+context("BigQuery Table")
 
-auth_mock = mock(cycle = T)
-insert_table_mock <- mock()
-with_mock(
-  `bigrquery::insert_table` = insert_table_mock,
-  `retl::bqAuth` = auth_mock,
-  test_that("Check that schema is passed to insert_table", {
+test_that("table is created from schema file", {
+  skip_on_travis()
+  res <- bqInitiateTable(
+    table = "test_table",
+    schema.file = "bq-table-schema.json",
+    partition = TRUE
+  )
 
-    bqInitiateTable(table = "test_table",
-                    schema.file = "bq-table-schema.json",
-                    partition = TRUE)
+  expected.schema <- list(
+    list(name = "id",
+         type = "INTEGER"),
+    list(name = "first_name",
+         type = "STRING")
+  )
+  expected.partition = TRUE
+})
 
-    expected.schema <- list(
-      list(name = "id",
-           type = "INTEGER"),
-      list(name = "first_name",
-           type = "STRING")
-    )
-    expected.partition = TRUE
-
-    expect_args(
-      insert_table_mock,
-      1,
-      project = Sys.getenv("BIGQUERY_PROJECT"),
-      dataset = Sys.getenv("BIGQUERY_DATASET"),
-      table = "test_table",
-      schema = expected.schema,
-      partition = expected.partition
-    )
-  })
-)